+++ src/api/routine/routine.controller.ts
@@ -1,6 +1,6 @@
 import type { Request, Response } from 'express';
 import { getUserStatistics, processDailyCheckin } from './routine.service.js';
-import { asyncHandler } from '../../core/asyncHandler.js';
+import { asyncHandler } from '../../middleware/asyncHandler.js';
 import { errorResponse, successResponse } from '../../core/response.js';
 
 export const dailyCheckinHandler = asyncHandler(async (req: Request, res: Response) => {
