+++ src/core/app.ts
@@ -2,7 +2,7 @@ import express, { type Application } from 'express';
 import { setupViewEngine } from './viewEngine.js';
 import welcomeRoutes from '../api/welcome/welcome.routes.js';
 import apiRoutes from '../routes/index.js';
-import { notFoundHandler } from './notFoundHandler.js';
+import { notFoundHandler } from '../middleware/notFoundHandler.js';
 import { errorHandler } from '../middleware/errorHandler.js';
 
 const app: Application = express();
