+++ src/middleware/notFoundHandler.ts
@@ -5,7 +5,6 @@ class HttpError extends Error {
   constructor(message: string, status: number) {
     super(message);
     this.status = status;
-    Object.setPrototypeOf(this, HttpError.prototype);
   }
 }
